Moved compiler flag setting to environment.sh
As it is necessary to have these compiler flags set, it seems more
straightfoward to me to have them set by default in the environment
script. This makes the setup commands slightly simpler and is less to
remember when recompiling. Devin recommends adding these commands to
your bashrc, which is reasonable, but I see no disadvantage in them
being set here as well.